(original rev 3637) merge dev into staging
Merged revisions 3557-3573,3583,3590,3606-3608,3611-3622,3630,3632,3634,3636 via svnmerge from
file:///gaba/svn/netgpu/omni/dev

................
  r3557 | mwimer | 2010-11-18 15:38:28 -0500 (Thu, 18 Nov 2010) | 2 lines

  alpha not darkening channel now
................
  r3558 | mwimer | 2010-11-19 16:42:08 -0500 (Fri, 19 Nov 2010) | 2 lines

  crosshairs and splitting crash on undo
................
  r3559 | mwimer | 2010-11-22 12:52:54 -0500 (Mon, 22 Nov 2010) | 2 lines

  only jump when validating
................
  r3560 | mwimer | 2010-11-22 15:12:23 -0500 (Mon, 22 Nov 2010) | 2 lines

  small change
................
  r3561 | mwimer | 2010-11-22 15:12:52 -0500 (Mon, 22 Nov 2010) | 2 lines

  part of the break threshold
................
  r3562 | mwimer | 2010-11-22 15:13:04 -0500 (Mon, 22 Nov 2010) | 2 lines

  break threshold
................
  r3563 | mwimer | 2010-11-22 15:13:40 -0500 (Mon, 22 Nov 2010) | 2 lines

  breaking based on threshold
................
  r3564 | mwimer | 2010-11-22 15:14:03 -0500 (Mon, 22 Nov 2010) | 2 lines

  use float not int for threshold
................
  r3565 | mwimer | 2010-11-22 15:14:34 -0500 (Mon, 22 Nov 2010) | 2 lines

  ibreak threshold
................
  r3566 | mwimer | 2010-11-22 15:35:21 -0500 (Mon, 22 Nov 2010) | 2 lines

  want to update as type
................
  r3567 | mwimer | 2010-11-22 15:39:37 -0500 (Mon, 22 Nov 2010) | 2 lines

  default to .8
................
  r3568 |  | 2010-11-22 16:12:56 -0500 (Mon, 22 Nov 2010) | 1 line

  rename function
................
  r3569 |  | 2010-11-22 18:11:03 -0500 (Mon, 22 Nov 2010) | 1 line

  start fiddling w/ drawing selected segment w/ highlight
................
  r3570 |  | 2010-11-22 18:25:14 -0500 (Mon, 22 Nov 2010) | 1 line

  center on searched-for segment
................
  r3571 |  | 2010-11-22 18:34:09 -0500 (Mon, 22 Nov 2010) | 1 line

  show color during break if segment was manually merged in
................
  r3572 | mwimer | 2010-11-23 12:12:51 -0500 (Tue, 23 Nov 2010) | 2 lines

  breaking in 2d as well now
................
  r3573 |  | 2010-11-24 02:32:01 -0500 (Wed, 24 Nov 2010) | 1 line

  add files for vegas mode, non-working version of omni.pro rewriter, and parrallel version of std::algorithms
................
  r3583 |  | 2010-11-24 03:52:47 -0500 (Wed, 24 Nov 2010) | 1 line

  pull aleks latest tree
................
  r3590 | mwimer | 2010-11-24 13:42:04 -0500 (Wed, 24 Nov 2010) | 2 lines

  qfes for autocenter
................
  r3606 | purcaro | 2010-11-26 23:43:59 -0500 (Fri, 26 Nov 2010) | 1 line

  guarentee that valid segments can never get joined when changing threshold
................
  r3607 | zlateski | 2010-11-27 15:48:37 -0500 (Sat, 27 Nov 2010) | 1 line

  watershed binary
................
  r3608 | zlateski | 2010-11-27 20:25:06 -0500 (Sat, 27 Nov 2010) | 1 line

  watershed source no sefgault
................
  r3611 | mwimer | 2010-11-29 14:30:20 -0500 (Mon, 29 Nov 2010) | 2 lines

  refuse to move items from container into itself since that fails
................
  r3612 | mwimer | 2010-11-29 14:52:13 -0500 (Mon, 29 Nov 2010) | 2 lines

  fixing split bug
................
  r3613 | mwimer | 2010-11-29 15:05:28 -0500 (Mon, 29 Nov 2010) | 2 lines

  uncertain...acting like not valid..both need to be refactored
................
  r3614 | mwimer | 2010-11-29 16:11:55 -0500 (Mon, 29 Nov 2010) | 2 lines

  making actions more ugly and ...yeah, hedious in there now.
................
  r3615 | mwimer | 2010-11-30 14:11:08 -0500 (Tue, 30 Nov 2010) | 2 lines

  hack to fix the depth bug
................
  r3616 | mwimer | 2010-11-30 15:19:47 -0500 (Tue, 30 Nov 2010) | 2 lines

  lumosity hack, not sure if people will use or like
................
  r3617 | mwimer | 2010-11-30 15:23:44 -0500 (Tue, 30 Nov 2010) | 3 lines

  style. haha
................
  r3618 | mwimer | 2010-12-01 14:03:46 -0500 (Wed, 01 Dec 2010) | 2 lines

  decrement
................
  r3619 |  | 2010-12-01 14:18:33 -0500 (Wed, 01 Dec 2010) | 1 line

  use int64 instaed of uint64
................
  r3620 | mwimer | 2010-12-01 15:07:11 -0500 (Wed, 01 Dec 2010) | 2 lines

  order of ops here means mvoxel needs to be added to before subed from if we want it to never be negative
................
  r3621 | mwimer | 2010-12-01 16:56:04 -0500 (Wed, 01 Dec 2010) | 2 lines

  jump based on different crieteria...basically if you are a solely selected segment then autocenter.
................
  r3622 | mwimer | 2010-12-01 17:07:02 -0500 (Wed, 01 Dec 2010) | 2 lines

  order
................
  r3630 | mwimer | 2010-12-02 14:33:36 -0500 (Thu, 02 Dec 2010) | 2 lines

  distance using better heuristic based on bbox
................
  r3632 | mwimer | 2010-12-02 16:15:48 -0500 (Thu, 02 Dec 2010) | 2 lines

  removing from groups now
................
  r3634 |  | 2010-12-02 16:20:56 -0500 (Thu, 02 Dec 2010) | 1 line

  start of fiddling w/ view2d prefs;
................
  r3636 |  | 2010-12-02 16:31:44 -0500 (Thu, 02 Dec 2010) | 16 lines

  Blocked revisions 2271,2275-2276 via svnmerge

  ........
    r2271 |  | 2010-07-25 00:11:43 -0400 (Sun, 25 Jul 2010) | 2 lines

    playing in experiemntal branch, now; import hdf5 into memmapped files...
  ........
    r2275 |  | 2010-07-25 00:22:52 -0400 (Sun, 25 Jul 2010) | 2 lines

    don't double fix the offset!
  ........
    r2276 |  | 2010-07-25 01:42:07 -0400 (Sun, 25 Jul 2010) | 2 lines

    hacing away
  ........
................


git-svn-id: file:///tmp/current/omni/branches/staging@3380 2ffc8166-1744-45da-a82a-d6ea10f7896b

+++ omni/src/viewGroup/omViewGroupState.h
@@ -58,8 +58,8 @@ public:
 											 const Vector2i&,
 											 const OmTileCoord&);
 
-	void setBreakThreshold(int t){ mBreakThreshold = t; }
-	int getBreakThreshold(){ return mBreakThreshold; }
+	void setBreakThreshold(float t){ mBreakThreshold = t; }
+	float getBreakThreshold(){ return mBreakThreshold; }
 
 	void setDustThreshold(unsigned int t){ mDustThreshold = t; }
 	unsigned int getDustThreshold(){ return mDustThreshold; }
@@ -97,7 +97,7 @@ private:
 
 	boost::shared_ptr<OmBrushSize> brushSize_;
 
-	int mBreakThreshold;
+	float mBreakThreshold;
 	unsigned int mDustThreshold;
 
 	//view event
